requirements.txt
Django==2.2.6
psycopg2-binary==2.8.3
djangorestframework==3.11.0